Cleanup some unused DTO's that were missing in merge Changes to transactions.

+++ src/NeoModules.RPC/TransactionManagers/ITransactionManager.cs
@@ -6,7 +6,7 @@ namespace NeoModules.RPC.TransactionManagers
     public interface ITransactionManager
     {
         IClient Client { get; set; }
-        Task<string> SendTransactionAsync(string from, string to, double amount);
+        Task<string> SendTransactionAsync(string from, string to, decimal amount);
         Task<bool> SendRawTransactionAsync(string hexTx);
     }
 }
